refactor(material-layer): Convert off of Next CSS-In-JS styles
This component will require serious review, a significant refactor was made to work around Next JS' CSS-in-JS and the logic is now rehoused in the script tag to be determined at runtime, declared in the render markdown dynamically, and then processed by rules outlined in the style tag. Reusing this feature may impact translation to other frameworks.